+++ tests/test_fasta_manager.py
@@ -1,3 +1,4 @@
+import copy
 import os
 from pathlib import Path
 
@@ -41,12 +42,26 @@ def test_download_ncbi(fasta_manager):
     assert content
 
 
+def test_download_ncbi_timeout(fasta_manager):
+    mngr = copy.deepcopy(fasta_manager)
+    mngr.timeout = 0.000001
+    with pytest.raises(RuntimeError):
+        mngr.download_ncbi(mite_acc="MITE9999999", genpept_acc="AAD28496.1")
+
+
 def test_download_uniprot(fasta_manager):
     path, content = fasta_manager.download_uniprot(
         mite_acc="MITE9999999", uniprot_acc="Q9X2V9"
     )
     assert isinstance(path, Path)
     assert content
+
+
+def test_download_uniprot_timeout(fasta_manager):
+    mngr = copy.deepcopy(fasta_manager)
+    mngr.timeout = 0.000001
+    with pytest.raises(RuntimeError):
+        mngr.download_uniprot(mite_acc="MITE9999999", uniprot_acc="Q9X2V9")
 
 
 def test_download_uniparc(fasta_manager):